+++ jpostman-junit/src/test/java/io/jpostman/junit/JUnitContextTest.java
@@ -68,7 +68,10 @@ public class JUnitContextTest {
 				.isEqual(cxt.asString("accessToken"), "abc123").isNotEqual(cxt.asString("accessToken"), "bad")
 				.isTrue(true).isFalse(false).isNull(null).isNotNull(cxt.response());
 
-		cxt.verify();
+		cxt.context(c -> cxt.verify());
+		cxt.soft().context().verify(400);
+		
+		cxt.soft(true).statusCode(400).exists("accessToken", "Access token not found").assertAll();
 
 		try {
 			cxt.soft(true).statusCode(400).exists("inactive").pathEquals("inactive", true)
@@ -206,15 +209,15 @@ public class JUnitContextTest {
 		JUnitContext secure = JUnitContext.create();
 		AtomicInteger calls = new AtomicInteger();
 
-		String first = secure.cache("token", () -> {
+		String first = secure.cache(() -> {
 			calls.incrementAndGet();
 			return "abc123";
-		});
+		}, "token");
 
-		String second = secure.cache("token", () -> {
+		String second = secure.cache(() -> {
 			calls.incrementAndGet();
 			return "new-value";
-		});
+		},"token");
 
 		String third = secure.cache(() -> {
 			calls.incrementAndGet();